% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_company_bubble.R
\name{prep_company_bubble}
\alias{prep_company_bubble}
\title{Prepare data for company bubble plot}
\usage{
prep_company_bubble(
  equity_results_company,
  bonds_results_company,
  portfolio_name,
  start_year,
  green_techs
)
}
\arguments{
\item{equity_results_company}{(data.frame) Company-level results from
\code{workflow.pacta} for listed equity}

\item{bonds_results_company}{(data.frame) Company-level results from
\code{workflow.pacta} for corporate bonds}

\item{portfolio_name}{(character scalar) Portfolio name}

\item{start_year}{(integer scalar) Start year for analysis}

\item{green_techs}{(character vector) List of green technologies}
}
\value{
(data.frame) suitible for serialization to JSON (using
\code{jsonlite::toJSON}/\code{jsonlite::write_json})
}
\description{
Prepare JSON data for the company bubble plot.
Note: this is based on pacta.portfolio.report:::prep_company_bubble, but
does not filter to allocation == "portfolio_weight" nor by scenario and
scenario source.
}
