% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_exposure_pie.R
\name{prep_exposure_pie}
\alias{prep_exposure_pie}
\title{Prepare data for exposure pie chart}
\usage{
prep_exposure_pie(
  data,
  asset_type,
  investor_name,
  portfolio_name,
  pacta_sectors,
  currency_exchange_value
)
}
\arguments{
\item{asset_type}{(character scalar) Asset type (\code{Equity} or \code{Bonds})}

\item{investor_name}{(character scalar) Investor name}

\item{portfolio_name}{(character scalar) Portfolio name}

\item{pacta_sectors}{(character vector) PACTA sectors}

\item{currency_exchange_value}{(numeric scalar) Currency exchange value to
convert to USD}

\item{(data.frame)}{Audit file contents}
}
\value{
(data.frame) suitible for serialization to JSON (using
\code{jsonlite::toJSON}/\code{jsonlite::write_json})
}
\description{
Prepare JSON data for the exposure pie chart.
}
