% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_techmix_sector.R
\name{prep_techmix_sector}
\alias{prep_techmix_sector}
\title{Prepare techmix data for plot}
\usage{
prep_techmix_sector(
  equity_results_portfolio,
  bonds_results_portfolio,
  indices_eq_results_portfolio,
  indices_cb_results_portfolio,
  peers_equity_results_portfolio,
  peers_bonds_results_portfolio,
  investor_name,
  portfolio_name,
  start_year,
  year_span,
  peer_group,
  green_techs,
  all_tech_levels
)
}
\arguments{
\item{equity_results_portfolio}{(data.frame) Portfolio-level results from
\code{workflow.pacta} for listed equity}

\item{bonds_results_portfolio}{(data.frame) Portfolio-level results from
\code{workflow.pacta} for corporate bonds}

\item{indices_eq_results_portfolio}{(data.frame) Benchmark results from
\code{workflow.pacta} for listed equity}

\item{indices_cb_results_portfolio}{(data.frame) Benchmark results from
\code{workflow.pacta} for corporate bonds}

\item{peers_equity_results_portfolio}{(data.frame) Peer group results from
\code{workflow.pacta} for listed equity}

\item{peers_bonds_results_portfolio}{(data.frame) Peer group results from
\code{workflow.pacta} for corporate bonds}

\item{investor_name}{(character scalar) Investor name}

\item{portfolio_name}{(character scalar) Portfolio name}

\item{start_year}{(integer scalar) Start year for analysis}

\item{year_span}{(integer scalar) Number of years to analyze (from
\verb{_start_year})}

\item{peer_group}{(character scalar) Peer group name}

\item{green_techs}{(character vector) List of green technologies}

\item{all_tech_levels}{(character vector) List of all technology levels}
}
\value{
(data.frame) suitible for serialization to JSON (using
\code{jsonlite::toJSON}/\code{jsonlite::write_json})
}
\description{
Prepare JSON data for the technology mix plot.
}
