% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dashboard_workflow.R
\name{run_dashboard_workflow}
\alias{run_dashboard_workflow}
\title{Run Dashboard workflow}
\usage{
run_dashboard_workflow(
  params,
  raw_params,
  run_analysis = TRUE,
  run_dashboard = TRUE,
  analysis_output_dir = Sys.getenv("ANALYSIS_OUTPUT_DIR"),
  benchmarks_dir = Sys.getenv("BENCHMARKS_DIR"),
  pacta_data_dir = Sys.getenv("PACTA_DATA_DIR"),
  portfolio_dir = Sys.getenv("PORTFOLIO_DIR"),
  dashboard_output_dir = Sys.getenv("DASHBOARD_OUTPUT_DIR")
)
}
\arguments{
\item{params}{(\code{list}) A list of parameters to be used in the analysis and
dashboard process. See JSON Schema for details.}

\item{raw_params}{(\code{character}) Raw JSON string of parameters.}

\item{run_analysis}{(\code{logical}) Run the analysis process.}

\item{run_dashboard}{(\code{logical}) Run the dashboard creation process.}

\item{analysis_output_dir}{Directory containing the PACTA analysis results.}

\item{benchmarks_dir}{filepath: Directory containing the benchmark analysis
results.}

\item{pacta_data_dir}{filepath: Directory with "pacta-data"}

\item{portfolio_dir}{filepath: Directory with portfolio files}

\item{dashboard_output_dir}{Directory where the dashboard will be
saved.}
}
\value{
(invisible) \code{TRUE} if the workflow was successful. Primarily called
for side effect of rendering files.
}
\description{
Run steps required to prepare a PACTA analysis and dashboard
}
