% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.pacta.dashboard-package.R
\docType{package}
\name{workflow.pacta.dashboard-package}
\alias{workflow.pacta.dashboard}
\alias{workflow.pacta.dashboard-package}
\title{workflow.pacta.dashboard: Run PACTA dashboard JSON generation}
\description{
Run PACTA dashboard JSON generation.
}
\author{
\strong{Maintainer}: Alex Axthelm \email{aaxthelm@rmi.org} (\href{https://orcid.org/0000-0001-8579-8565}{ORCID}) [contractor]

Authors:
\itemize{
  \item CJ Yetman \email{cj@cjyetman.com} (\href{https://orcid.org/0000-0001-5099-9500}{ORCID}) [contractor]
  \item Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID}) [contractor]
}

Other contributors:
\itemize{
  \item RMI \email{PACTA4investors@rmi.org} [copyright holder, funder]
}

}
\keyword{internal}
